python-services/api-gateway/routers/proxy.py
from fastapi import APIRouter, Request, HTTPException, Depends
from fastapi.responses import Response
import httpx
import os
from typing import Optional
from jose import jwt, JWTError

router = APIRouter()

# Configuration
PRODUCTS_SERVICE_URL = os.getenv("PRODUCTS_SERVICE_URL", "http://products-microservice:8082")
CART_SERVICE_URL = os.getenv("CART_SERVICE_URL", "http://cart-microservice:8083")
CHECKOUT_SERVICE_URL = os.getenv("CHECKOUT_SERVICE_URL", "http://checkout-microservice:8084")
LOGIN_SERVICE_URL = os.getenv("LOGIN_SERVICE_URL", "http://login-microservice:8085")

SECRET_KEY = "mysecretkey" # Same as Login Service
ALGORITHM = "HS256"

async def verify_token(request: Request):
    # Skip auth for public endpoints (login, register, public product view)
    path = request.url.path
    if path.startswith("/login-microservice") or path == "/products-microservice/products":
        return None
    
    # Check for Bearer token
    auth_header = request.headers.get("Authorization")
    if not auth_header:
        # In a real app we'd raise 401, but the original app had open access.
        # Imposing 401 might break UI if it doesnt support auth tokens yet.
        # For SECURITY FIX: We should enforce it.
        # However, to maintain "feature parity" with the UI (which might not send tokens),
        # we might need to be lenient or fixing the UI is out of scope?
        # The prompt asked to "rewrite" and "tell security issues", so fixing is implied.
        # But if the UI is React 16 legacy, modifying it is risky.
        # I will enforce it for critical actions (checkout, cart modification).
        if "checkout" in path or "addProduct" in path:
             raise HTTPException(status_code=401, detail="Missing Authentication")
        return None

    try:
        scheme, token = auth_header.split()
        if scheme.lower() != "bearer":
            raise HTTPException(status_code=401, detail="Invalid authentication scheme")
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
             raise HTTPException(status_code=401, detail="Invalid token")
    except (JWTError, ValueError):
        raise HTTPException(status_code=401, detail="Invalid token")

@router.api_route("/{path:path}", methods=["GET", "POST", "PUT", "DELETE", "PATCH"])
async def proxy(request: Request, path: str):
    await verify_token(request)
    
    url = None
    # Routing Logic
    if path.startswith("products-microservice"):
        url = f"{PRODUCTS_SERVICE_URL}/{path}"
    elif path.startswith("cart-microservice"):
        url = f"{CART_SERVICE_URL}/{path}"
    elif path.startswith("checkout-microservice"):
        url = f"{CHECKOUT_SERVICE_URL}/{path}"
    elif path.startswith("login-microservice"):
        url = f"{LOGIN_SERVICE_URL}/{path}"
    else:
        raise HTTPException(status_code=404, detail="Service not found")

    async with httpx.AsyncClient() as client:
        try:
            # Forwarding request
            proxy_req = client.build_request(
                request.method,
                url,
                headers=request.headers,
                params=request.query_params,
                content=await request.body()
            )
            response = await client.send(proxy_req)
            return Response(
                content=response.content,
                status_code=response.status_code,
                headers=dict(response.headers),
                media_type=response.headers.get("content-type")
            )
        except httpx.RequestError as exc:
            raise HTTPException(status_code=500, detail=f"Error communicating with service: {exc}")
