python-services/api-gateway/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
py-eureka-client>=0.10.4
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
